backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
dnspython==2.7.0
docker==7.1.0
drf-yasg==1.21.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
packaging==24.2
paramiko==3.5.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pymongo==4.11.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2
social-auth-app-django==5.4.2
